Refactor: problem form fix out of async processor
The "problem" form handling is not a key concern of the async form processor, so it makes sense to move it out to keep it simple. This is a refactor in preparation to handle more "problem" forms in the same or a similar way.
